pyproject.toml
[project]
name = "mism-exec"
version = "0.1.0"
description = "MISM Execution Platform — orchestrates model execution on Kubernetes"
requires-python = ">=3.10"
license = "Apache-2.0"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "kubernetes>=28.0",
    "jinja2>=3.1",
    "pydantic-settings>=2.2",
    "httpx>=0.27",
    "mism-registry[postgres] @ git+https://github.com/mism-center/metadata-schema.git@feature/mism-81",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "src/api",
    "src/core",
    "src/middleware",
    "src/orchestration",
    "src/schemas",
    "src/services",
]
sources = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort (import ordering)
    "N",   # pep8-naming
    "UP",  # pyupgrade (modernise syntax)
    "B",   # flake8-bugbear (common bugs)
    "SIM", # flake8-simplify
]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
